clened up accounts controller
