+++ app/controllers/accounts_controller.rb
@@ -1,6 +1,5 @@
 class AccountsController < ApplicationController
-  rescue_from ActiveRecord::RecordNotFound, with: :handle_not_found_account
-  rescue_from Exceptions::LowAccountBalanceError, with: :handle_low_balance_error
+  include RescueableAccountErrors
 
   def send_money
     @debit_account = Account.verified_and_fetch_by_email_or_phone(account_params[:debitor])
@@ -12,13 +11,5 @@ class AccountsController < ApplicationController
   private
   def account_params
     params.require(:account).permit(:amount, creditor: [:email, :phone], debitor: [:email, :phone])
-  end
-
-  def handle_not_found_account(exception)
-    render json: { error: exception.message }, status: :forbidden
-  end
-
-  def handle_low_balance_error(exception)
-    render json: { error: exception.message }, status: :forbidden
   end
 end
