--- spec/examples.txt
@@ -1,21 +1,0 @@
-example_id                                              | status | run_time        |
-------------------------------------------------------- | ------ | --------------- |
-./spec/controllers/accounts_controller_spec.rb[1:1:1:1] | passed | 0.11995 seconds |
-./spec/controllers/accounts_controller_spec.rb[1:1:1:2] | passed | 0.00992 seconds |
-./spec/models/account_spec.rb[1:1]                      | passed | 0.63799 seconds |
-./spec/models/account_spec.rb[1:2:1]                    | passed | 0.00959 seconds |
-./spec/models/account_spec.rb[1:2:2]                    | passed | 0.00956 seconds |
-./spec/models/account_spec.rb[1:2:3]                    | passed | 0.01437 seconds |
-./spec/models/account_spec.rb[1:2:4]                    | passed | 0.01072 seconds |
-./spec/models/account_spec.rb[1:3:1]                    | passed | 0.01529 seconds |
-./spec/models/account_spec.rb[1:3:2]                    | passed | 0.01644 seconds |
-./spec/models/account_spec.rb[1:3:3]                    | passed | 0.02374 seconds |
-./spec/models/account_spec.rb[1:4:1]                    | passed | 0.02 seconds    |
-./spec/models/account_spec.rb[1:4:2]                    | passed | 0.03093 seconds |
-./spec/models/account_spec.rb[1:4:3]                    | passed | 0.03938 seconds |
-./spec/requests/accounts_spec.rb[1:1:1]                 | passed | 0.04844 seconds |
-./spec/requests/accounts_spec.rb[1:1:2]                 | passed | 0.03929 seconds |
-./spec/requests/accounts_spec.rb[1:1:3]                 | passed | 0.03903 seconds |
-./spec/requests/accounts_spec.rb[1:1:4]                 | passed | 0.04162 seconds |
-./spec/requests/accounts_spec.rb[1:1:5]                 | passed | 0.05226 seconds |
-./spec/requests/accounts_spec.rb[1:1:6]                 | passed | 0.04204 seconds |
